+++ sdk/core/core-http/test/policies/tracingPolicyTests.ts
@@ -1,28 +1,28 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { assert } from "chai";
 import {
-  RequestPolicy,
-  WebResource,
-  HttpOperationResponse,
   HttpHeaders,
-  RequestPolicyOptions
+  HttpOperationResponse,
+  RequestPolicy,
+  RequestPolicyOptions,
+  WebResource
 } from "../../src/coreHttp";
+import { Span, SpanOptions, Tracer, TracerProvider, trace } from "@opentelemetry/api";
 import {
+  SpanAttributeValue,
+  SpanAttributes,
   SpanContext,
+  SpanStatus,
+  SpanStatusCode,
   TraceFlags,
   TraceState,
-  setSpan,
   context,
-  SpanStatusCode,
-  SpanStatus,
-  SpanAttributes,
-  SpanAttributeValue
+  setSpan
 } from "@azure/core-tracing";
-import { tracingPolicy } from "../../src/policies/tracingPolicy";
-import { TracerProvider, Tracer, Span, SpanOptions, trace } from "@opentelemetry/api";
+import { assert } from "chai";
 import sinon from "sinon";
+import { tracingPolicy } from "../../src/policies/tracingPolicy";
 
 class MockSpan implements Span {
   private _endCalled = false;