+++ sdk/mixedreality/mixed-reality-authentication/src/util/jwt.ts
@@ -7,7 +7,7 @@ import { base64decode } from "./base64";
  * Retrieves the expiration value from the JWT.
  *
  * @internal
- * @param jwtValue The JWT value.
+ * @param jwtValue - The JWT value.
  */
 export function retrieveJwtExpirationTimestamp(jwtValue: string): number {
   // Note: This only supports JWS Compact Serialization format.'