+++ sdk/sql/arm-sql/src/operations/serverAdvisors.ts
@@ -10,7 +10,7 @@ import { ServerAdvisors } from "../operationsInterfaces";
 import * as coreClient from "@azure/core-client";
 import * as Mappers from "../models/mappers";
 import * as Parameters from "../models/parameters";
-import { SqlManagementClientContext } from "../sqlManagementClientContext";
+import { SqlManagementClient } from "../sqlManagementClient";
 import {
   ServerAdvisorsListByServerOptionalParams,
   ServerAdvisorsListByServerResponse,
@@ -23,13 +23,13 @@ import {
 
 /** Class containing ServerAdvisors operations. */
 export class ServerAdvisorsImpl implements ServerAdvisors {
-  private readonly client: SqlManagementClientContext;
+  private readonly client: SqlManagementClient;
 
   /**
    * Initialize a new instance of the class ServerAdvisors class.
    * @param client Reference to the service client
    */
-  constructor(client: SqlManagementClientContext) {
+  constructor(client: SqlManagementClient) {
     this.client = client;
   }
 
@@ -152,7 +152,7 @@ const updateOperationSpec: coreClient.OperationSpec = {
     },
     default: {}
   },
-  requestBody: Parameters.parameters18,
+  requestBody: Parameters.parameters16,
   queryParameters: [Parameters.apiVersion2],
   urlParameters: [
     Parameters.$host,